2019.02.16 - refactor signupJson
