+++ src/main/java/com/mashup6th/preambackend/service/UserService.java
@@ -1,6 +1,6 @@
 package com.mashup6th.preambackend.service;
 
-import com.mashup6th.preambackend.dto.SignUpJson;
+import com.mashup6th.preambackend.dto.user.SignUpJson;
 import com.mashup6th.preambackend.dto.user.UserLoginInfo;
 import com.mashup6th.preambackend.entity.User;
