+++ src/main/java/com/mashup6th/preambackend/service/UserServiceImpl.java
@@ -1,7 +1,7 @@
 package com.mashup6th.preambackend.service;
 
 
-import com.mashup6th.preambackend.dto.SignUpJson;
+import com.mashup6th.preambackend.dto.user.SignUpJson;
 import com.mashup6th.preambackend.dto.user.UserLoginInfo;
 import com.mashup6th.preambackend.entity.User;
 import com.mashup6th.preambackend.entity.constant.AuthNumber;
